fix(checker): split symbol type depth guard
