+++ crates/tsz-checker/src/state/type_resolution/mod.rs
@@ -15,6 +15,7 @@ pub(crate) mod reference_helpers;
 pub(crate) mod reference_type_params;
 pub(crate) mod shadowed_lib_heritage;
 pub(crate) mod symbol_types;
+mod symbol_types_depth;
 pub(crate) mod symbol_types_dynamic_alias;
 pub(crate) mod symbol_types_lazy;
 #[cfg(test)]
