+++ crates/tsz-checker/src/state/type_resolution/symbol_types.rs
@@ -4,6 +4,7 @@
 use crate::query_boundaries::state::type_resolution as query;
 use crate::query_boundaries::type_predicates::is_compiler_managed_type;
 use crate::state::CheckerState;
+use crate::state_domain::type_resolution::symbol_types_depth::TypeReferenceResolutionDepthGuard;
 use crate::symbols_domain::alias_cycle::AliasCycleTracker;
 use crate::symbols_domain::name_text::entity_name_text_in_arena;
 use crate::types_domain::queries::lib_resolution::resolve_name_to_lib_symbol;
@@ -12,42 +13,6 @@ use tsz_parser::parser::node::{NodeAccess, NodeArena};
 use tsz_parser::parser::{NodeIndex, syntax_kind_ext};
 use tsz_scanner::SyntaxKind;
 use tsz_solver::TypeId;
-
-/// Maximum nesting depth for `type_reference_symbol_type_with_params`'s
-/// alias-forwarding recursion. A mutually-aliasing pair produced by a
-/// raw-`SymbolId` cross-file collision (`Dataset` ↔ `OutputDataset`) would
-/// otherwise ping-pong through the recursion until the stack overflows and
-/// aborts the compile. The cap is far above any legitimate alias chain and far
-/// below stack exhaustion, so valid code is unaffected. Refs #13212.
-const MAX_TYPE_REFERENCE_RESOLUTION_DEPTH: u32 = 350;
-
-thread_local! {
-    static TYPE_REFERENCE_RESOLUTION_DEPTH: std::cell::Cell<u32> = const { std::cell::Cell::new(0) };
-}
-
-/// RAII depth counter for
-/// [`CheckerState::type_reference_symbol_type_with_params`].
-struct TypeReferenceResolutionDepthGuard;
-
-impl TypeReferenceResolutionDepthGuard {
-    /// Enters one recursion level; returns `None` once the depth cap is hit.
-    fn enter() -> Option<Self> {
-        TYPE_REFERENCE_RESOLUTION_DEPTH.with(|depth| {
-            if depth.get() >= MAX_TYPE_REFERENCE_RESOLUTION_DEPTH {
-                None
-            } else {
-                depth.set(depth.get() + 1);
-                Some(Self)
-            }
-        })
-    }
-}
-
-impl Drop for TypeReferenceResolutionDepthGuard {
-    fn drop(&mut self) {
-        TYPE_REFERENCE_RESOLUTION_DEPTH.with(|depth| depth.set(depth.get().saturating_sub(1)));
-    }
-}
 
 impl<'a> CheckerState<'a> {
     pub(crate) fn type_reference_symbol_type(&mut self, sym_id: SymbolId) -> TypeId {
